Move tests to Old directory
