+++ .gitignore
@@ -5,4 +5,4 @@ vendor/
 composer.phar
 runtests.sh
 /ssh.*
-/tests/data/out/tables/
+/tests/Old/data/out/tables/
