+++ docker-compose.yml
@@ -31,7 +31,7 @@ services:
       MYSQL_ROOT_PASSWORD: somePassword
 
   sshproxy:
-    build: ./tests/env/sshproxy
+    build: ./tests/Old/env/sshproxy
     command: sh -c 'echo $SSH_KEY_PUBLIC >> ~/.ssh/authorized_keys && /usr/sbin/sshd -D'
     environment:
       - SSH_KEY_PUBLIC
